pyproject.toml
[project]
name = "glacium"
dynamic = ["version"]
description = ""
authors = [
    {name = "pineOapple", email = "44525727+pineOapple@users.noreply.github.com"}
]
readme = "README.md"
license = "MIT"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.12"
dependencies = [
    "click>=8",
    "pyyaml>=6",
    "rich>=13",
    "jinja2>=3",
    "coloredlogs>=15",
    "verboselogs>=1",
    "numpy>=1",
    "matplotlib>=3",
    "fpdf2 (>=2.8.3,<3.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0",
    "poetry-dynamic-versioning>=1.0.0,<2.0.0",
    "setuptools_scm>=7"]
build-backend = "poetry_dynamic_versioning.backend"

[project.scripts]
glacium = "glacium.cli:cli"

[tool.poetry]
exclude = [
    ".idea",
    "glacium/.idea",
    "runs",
    "docs/_build",
]
include = [
    "glacium/DejaVuSans.ttf",
]
version = "0.0.0"

[tool.poetry.group.dev.dependencies]
sphinx = "^8.2.3"
myst-parser = "^4.0.1"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "pep440"
local = false
